Moved subroutine that warns about .pde files in feedback package
