+++ internal/cli/compile/compile.go
@@ -158,12 +158,12 @@ func runCompileCommand(cmd *cobra.Command, args []string) {
 		path = args[0]
 	}
 
-	sketchPath := arguments.InitSketchPath(path, true)
-
+	sketchPath := arguments.InitSketchPath(path)
 	sk, err := sketch.LoadSketch(context.Background(), &rpc.LoadSketchRequest{SketchPath: sketchPath.String()})
 	if err != nil {
 		feedback.FatalError(err, feedback.ErrGeneric)
 	}
+	feedback.WarnAboutDeprecatedFiles(sk)
 
 	var inst *rpc.Instance
 	var profile *rpc.Profile
